Change to have equals and gethashcode in base classes
