+++ src/FlowtideDotNet.Substrait/Relations/BufferRelation.cs
@@ -39,7 +39,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is BufferRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                     Equals(Input, relation.Input);
             }
             return false;
@@ -48,13 +48,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Input);
             return code.ToHashCode();
         }
