+++ src/FlowtideDotNet.Substrait/Relations/FetchRelation.cs
@@ -48,7 +48,7 @@ namespace FlowtideDotNet.Substrait.Relations
             if (obj is FetchRelation relation)
             {
                 
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                     Equals(Input, relation.Input) &&
                    Equals(Count, relation.Count) &&
                    Equals(Offset, relation.Offset);
@@ -59,13 +59,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Input);
             code.Add(Offset);
             code.Add(Count);
