+++ src/FlowtideDotNet.Substrait/Relations/IterationRelation.cs
@@ -46,7 +46,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is IterationRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(Input, relation.Input) &&
                    Equals(IterationName, relation.IterationName) &&
                    Equals(LoopPlan, relation.LoopPlan) &&
@@ -59,13 +59,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             if (Input != null)
             {
                 code.Add(Input);
