+++ src/FlowtideDotNet.Substrait/Relations/MergeJoinRelation.cs
@@ -49,7 +49,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is MergeJoinRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(Type, relation.Type) &&
                    Equals(Left, relation.Left) &&
                    Equals(Right, relation.Right) &&
@@ -63,13 +63,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Type);
             code.Add(Left);
             code.Add(Right);
