+++ src/FlowtideDotNet.Substrait/Relations/ProjectRelation.cs
@@ -42,7 +42,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is ProjectRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(Input, relation.Input) &&
                    Expressions.SequenceEqual(relation.Expressions);
             }
@@ -52,13 +52,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Input);
             foreach(var expression in Expressions)
             {
