+++ src/FlowtideDotNet.Substrait/Relations/ReadRelation.cs
@@ -44,7 +44,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is ReadRelation relation)
             {
-                return EmitEquals(relation.Emit) && 
+                return base.Equals(relation) && 
                    Equals(BaseSchema, relation.BaseSchema) &&
                    Equals(NamedTable, relation.NamedTable) &&
                    Equals(Filter, relation.Filter);
@@ -55,13 +55,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(BaseSchema);
             code.Add(NamedTable);
             code.Add(Filter);
