+++ src/FlowtideDotNet.Substrait/Relations/RootRelation.cs
@@ -30,7 +30,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is RootRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(Input, relation.Input) &&
                    Names.SequenceEqual(relation.Names);
             }
@@ -40,13 +40,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Input);
             foreach(var name in Names)
             {
