+++ src/FlowtideDotNet.Substrait/Relations/SetRelation.cs
@@ -40,7 +40,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is SetRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Inputs.SequenceEqual(relation.Inputs) &&
                    Equals(Operation, relation.Operation);
             }
@@ -50,13 +50,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             foreach (var input in Inputs)
             {
                 code.Add(input);
