+++ src/FlowtideDotNet.Substrait/Relations/SortRelation.cs
@@ -46,7 +46,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is SortRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(Input, relation.Input) &&
                    Sorts.SequenceEqual(relation.Sorts);
             }
@@ -56,13 +56,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Input);
             foreach (var sort in Sorts)
             {
