+++ src/FlowtideDotNet.Substrait/Relations/TableFunctionRelation.cs
@@ -55,7 +55,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is TableFunctionRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(TableFunction, relation.TableFunction) &&
                    Equals(Input, relation.Input) &&
                    Equals(Type, relation.Type) &&
@@ -67,13 +67,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(TableFunction);
             code.Add(Input);
             code.Add(Type);
