+++ src/FlowtideDotNet.Substrait/Relations/TopNRelation.cs
@@ -50,7 +50,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is TopNRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(Input, relation.Input) &&
                    Sorts.SequenceEqual(relation.Sorts) &&
                    Equals(Offset, relation.Offset) &&
@@ -62,13 +62,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Input);
             foreach (var sort in Sorts)
             {
