+++ src/FlowtideDotNet.Substrait/Relations/UnwrapRelation.cs
@@ -50,7 +50,7 @@ namespace FlowtideDotNet.Substrait.Relations
         {
             if (obj is UnwrapRelation relation)
             {
-                return EmitEquals(relation.Emit) &&
+                return base.Equals(relation) &&
                    Equals(Input, relation.Input) &&
                    Equals(BaseSchema, relation.BaseSchema) &&
                    Equals(Filter, relation.Filter) &&
@@ -62,13 +62,7 @@ namespace FlowtideDotNet.Substrait.Relations
         public override int GetHashCode()
         {
             var code = new HashCode();
-            if (Emit != null)
-            {
-                foreach (var emit in Emit)
-                {
-                    code.Add(emit);
-                }
-            }
+            code.Add(base.GetHashCode());
             code.Add(Input);
             code.Add(BaseSchema);
             code.Add(Filter);
