+++ src/FlowtideDotNet.Substrait/Type/AnyType.cs
@@ -16,16 +16,5 @@ namespace FlowtideDotNet.Substrait.Type
     public class AnyType : SubstraitBaseType
     {
         public override SubstraitType Type => SubstraitType.Any;
-
-        public override bool Equals(object? obj)
-        {
-            return obj is AnyType type &&
-                   Nullable == type.Nullable;
-        }
-
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(Type, Nullable);
-        }
     }
 }
