+++ src/FlowtideDotNet.Substrait/Type/DateType.cs
@@ -16,16 +16,5 @@ namespace FlowtideDotNet.Substrait.Type
     public class DateType : SubstraitBaseType
     {
         public override SubstraitType Type => SubstraitType.Date;
-
-        public override bool Equals(object? obj)
-        {
-            return obj is DateType type &&
-                   Nullable == type.Nullable;
-        }
-
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(Type, Nullable);
-        }
     }
 }
